Add support for duration metrics
