+++ main.go
@@ -3,7 +3,7 @@ package main
 import (
 	"context"
 	"fmt"
-	"github.com/fnproject/ext-metrics/handlers"
+	"github.com/fnproject/ext-metrics/statistics"
 	"github.com/fnproject/fn/api/server"
 )
 
@@ -12,7 +12,7 @@ func main() {
 	funcServer := server.NewFromEnv(ctx)
 
 	// Add the metrics API extension
-	handlers.AddEndpoints(funcServer)
+	statistics.AddEndpoints(funcServer)
 
 	fmt.Println("Starting fn server with metrics API extensions")
 	funcServer.Start(ctx)
