+++ prometheus.yml
@@ -1,0 +1,20 @@
+global:
+  scrape_interval:     15s # By default, scrape targets every 15 seconds.
+
+  # Attach these labels to any time series or alerts when communicating with
+  # external systems (federation, remote storage, Alertmanager).
+  external_labels:
+    monitor: 'fn-monitor'
+
+# A scrape configuration containing exactly one endpoint to scrape:
+# Here it's the Fn server
+scrape_configs:
+  # The job name is added as a label `job=<job_name>` to any timeseries scraped from this config.
+  - job_name: 'functions'
+
+    # Override the global default and scrape targets from this job every 5 seconds.
+    scrape_interval: 5s
+
+    static_configs:
+      # Specify all the fn servers from which metrics will be scraped
+      - targets: ['fnserver:8080'] # Uses /metrics by default
